Move few PackageListing serializers to shared files
Move EmptyStringAsNoneField and CyberstormPackageTeamSerializer to shared serializers files, so that they can be cleanly used by PackageVersionAPIView
